rag_poc/requirements.txt
fastapi
uvicorn
openai==0.28
pymilvus==2.2.17
python-dotenv
tqdm
sentence-transformers